langchain/conversational_app/requirements.txt
langchain
langchain_community
langchain_openai
huggingface_hub
python-dotenv
openai
pypdf
faiss-cpu
google-search-results